requirements.txt
fastapi[standard]
uvicorn
chromadb
sentence-transformers
pdfplumber
python-docx
groq
pydantic-settings
python-multipart
python-dotenv
faster-whisper
TTS
numpy
livekit
